Layers and weird syncing issues resolved.
