+++ deployment/configurations/node/sage3-prod.hjson
@@ -112,7 +112,7 @@
     "apps": [
       "Chat",
       "ChartMaker",
-      "MapGL",
+      "Map",
       "Notepad",
       "Poll",
       "SageCell",
