+++ webstack/libs/applications/src/lib/apps.json
@@ -1,41 +1,41 @@
 [
-    "AIPane",
-    "AssetLink",
-    "BoardLink",
-    "CSVViewer",
-    "Calculator",
-    "ChartGenerator",
-    "Chat",
-    "Clock",
-    "Cobrowse",
-    "CodeEditor",
-    "Counter",
-    "DeepZoomImage",
-    "EChartsViewer",
-    "GLTFViewer",
-    "HCDP",
-    "IFrame",
-    "ImageViewer",
-    "JupyterLab",
-    "LeafLet",
-    "Linker",
-    "MapGL",
-    "Notepad",
-    "PDFResult",
-    "PDFViewer",
-    "PluginApp",
-    "Poll",
-    "RTCChat",
-    "SageCell",
-    "Seer",
-    "SensorOverview",
-    "Stickie",
-    "TLDraw",
-    "Timer",
-    "TwilioScreenshare",
-    "VegaLite",
-    "VegaLiteViewer",
-    "VideoViewer",
-    "WebpageLink",
-    "Webview"
+  "AIPane",
+  "AssetLink",
+  "BoardLink",
+  "CSVViewer",
+  "Calculator",
+  "ChartGenerator",
+  "Chat",
+  "Clock",
+  "Cobrowse",
+  "CodeEditor",
+  "Counter",
+  "DeepZoomImage",
+  "EChartsViewer",
+  "GLTFViewer",
+  "HCDP",
+  "IFrame",
+  "ImageViewer",
+  "JupyterLab",
+  "LeafLet",
+  "Linker",
+  "Map",
+  "Notepad",
+  "PDFResult",
+  "PDFViewer",
+  "PluginApp",
+  "Poll",
+  "RTCChat",
+  "SageCell",
+  "Seer",
+  "SensorOverview",
+  "Stickie",
+  "TLDraw",
+  "Timer",
+  "TwilioScreenshare",
+  "VegaLite",
+  "VegaLiteViewer",
+  "VideoViewer",
+  "WebpageLink",
+  "Webview"
 ]
