+++ webstack/libs/applications/src/lib/apps/Map/index.ts
@@ -6,7 +6,20 @@
  * the file LICENSE, distributed as part of this software.
  */
 
+import { colors } from '@sage3/shared';
 import { z } from 'zod';
+
+const zColors = z.enum([...colors]);
+const colorScale = z.enum(['greys', 'inferno', 'viridis', 'turbo']);
+
+const Layer = z.object({
+  assetId: z.string(),
+  visible: z.boolean(),
+  color: zColors,
+  colorScale: colorScale,
+  opacity: z.number().min(0).max(1),
+});
+export type LayerType = z.infer<typeof Layer>;
 
 /**
  * SAGE3 application: MapGL
@@ -19,16 +32,9 @@ export const schema = z.object({
   bearing: z.number(),
   pitch: z.number(),
   baseLayer: z.string(),
-  overlay: z.boolean(),
   assetid: z.string().optional(),
   // Remove single `assetid`; instead, keep an array of layers:
-  layers: z.array(
-    z.object({
-      assetid: z.string(),
-      visible: z.boolean(),
-    })
-  ),
-  colorScale: z.string(),
+  layers: z.array(Layer),
 });
 export type state = z.infer<typeof schema>;
 
@@ -38,11 +44,7 @@ export const init: Partial<state> = {
   bearing: 0,
   pitch: 0,
   baseLayer: 'OpenStreetMap',
-  overlay: true,
-  assetid: '',
-  // Start with no overlay layers
   layers: [],
-  colorScale: 'greys',
 };
 
-export const name = 'MapGL';
+export const name = 'Map';
