+++ webstack/libs/frontend/src/lib/hooks/useFiles.tsx
@@ -108,7 +108,14 @@ async function openApplication(a: Asset, xDrop: number, yDrop: number, roomId: s
   } else {
     // Check all the supported file types
     if (isGeoTiff(fileType)) {
-      return setupApp(a.data.originalfilename, 'MapGL', xDrop, yDrop, roomId, boardId, { w: w, h: w }, { assetid: fileID });
+      const initialLayer = {
+        assetId: fileID,
+        visible: true,
+        color: 'red',
+        colorScale: 'turbo',
+        opacity: 1,
+      } as NonNullable<(typeof initialValues)['Map']['layers']>[0];
+      return setupApp(a.data.originalfilename, 'Map', xDrop, yDrop, roomId, boardId, { w: w, h: w }, { layers: [initialLayer] });
     } else if (isFileURL(fileType)) {
       const localurl = apiUrls.assets.getAssetById(a.data.file);
       // Get the content of the file
@@ -176,7 +183,14 @@ async function openApplication(a: Asset, xDrop: number, yDrop: number, roomId: s
           const metadata = await response.json();
           // Check if it is a GeoTiff
           if (metadata && metadata.GeoTiffVersion) {
-            return setupApp(a.data.originalfilename, 'MapGL', xDrop, yDrop, roomId, boardId, { w: w, h: w }, { assetid: fileID });
+            const initialLayer = {
+              assetId: fileID,
+              visible: true,
+              color: 'red',
+              colorScale: 'turbo',
+              opacity: 1,
+            } as NonNullable<(typeof initialValues)['Map']['layers']>[0];
+            return setupApp(a.data.originalfilename, 'Map', xDrop, yDrop, roomId, boardId, { w: w, h: w }, { layers: [initialLayer] });
           }
         }
       }
@@ -211,7 +225,14 @@ async function openApplication(a: Asset, xDrop: number, yDrop: number, roomId: s
     } else if (isGLTF(fileType)) {
       return setupApp('', 'GLTFViewer', xDrop, yDrop, roomId, boardId, { w: 600, h: 600 }, { assetid: fileID });
     } else if (isGeoJSON(fileType)) {
-      return setupApp('', 'MapGL', xDrop, yDrop, roomId, boardId, { w: 800, h: 400 }, { assetid: fileID });
+      const initialLayer = {
+        assetId: fileID,
+        visible: true,
+        color: 'red',
+        colorScale: 'turbo',
+        opacity: 1,
+      } as NonNullable<(typeof initialValues)['Map']['layers']>[0];
+      return setupApp('', 'Map', xDrop, yDrop, roomId, boardId, { w: 800, h: 400 }, { layers: [initialLayer] });
     } else if (isMD(fileType)) {
       const localurl = apiUrls.assets.getAssetById(a.data.file);
       // Get the content of the file
