+++ webstack/sage3-prod.hjson
@@ -110,7 +110,7 @@
     "apps": [
       "Chat",
       "ChartMaker",
-      "MapGL",
+      "Map",
       "Notepad",
       "Poll",
       "SageCell",
